rename app folder to src, delete button component and create store folder and move actions and reducers inside
