--- App/Components/Button/Button.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-class Button extends React.Component {
-  render(){
-    return(
-      <View></View
-    );
-  }
-}
-
-export default Button;
