--- App/Components/Button/index.js
@@ -1,5 +1,0 @@
-import styles from './styles';
-import Button from './Button';
-
-
-export {Button, styles};
