--- App/Components/Button/styles.js
@@ -1,7 +1,0 @@
-import EStyleSheet from 'react-native-extended-stylesheet';
-
-const styles = EStyleSheet.create({
-
-});
-
-export default styles;
