+++ src/components/Logo/styles.js
